# [Enhancement] faster but nondeterministic version of hard voxelization