--- a/tests/test_models/test_voxel_encoder/test_voxelize.py
+++ b/tests/test_models/test_voxel_encoder/test_voxelize.py
@@ -85,6 +85,9 @@ def test_voxelization():
 
 
 def test_voxelization_nondeterministic():
+    if not torch.cuda.is_available():
+        pytest.skip('test requires GPU and torch+cuda')
+
     voxel_size = [0.5, 0.5, 0.5]
     point_cloud_range = [0, -40, -3, 70.4, 40, 1]
     data_path = './tests/data/kitti/training/velodyne_reduced/000000.bin'