Small enhancement -- reimplement hard voxelization and make it faster at the cost of deterministic property

It is now an optional feature and is disabled by default. Set `deterministic=False` inside `Voxelization` config to enable it. It only works for hard voxelization and has no effect when using dynamic voxelization.

Performance comparison:
| conf | training iter time tested on | iter time before| iter time after|
|:-------:|:-----------------------------------------:|:----------------------:|:---------------------:|
| [pointpillars-kitti-3class](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/pointpillars/hv_pointpillars_secfpn_6x8_160e_kitti-3d-3class.py) | 2080Ti<br>bs=6 | 0.340 | 0.303 |
| [pointpillars-kitti-car](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/pointpillars/hv_pointpillars_secfpn_6x8_160e_kitti-3d-car.py) | 2080Ti<br>bs=6 | 0.281 | 0.246 |
| [pointpillars-waymo-3class](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/pointpillars/hv_pointpillars_secfpn_sbn_2x16_2x_waymo-3d-3class.py) | 3090<br>bs=2 | 0.412 | 0.296 |
| [pointpillars-waymo-car](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/pointpillars/hv_pointpillars_secfpn_sbn_2x16_2x_waymo-3d-car.py) | 3090<br>bs=2 | 0.365 | 0.237 |
| [centerpoint-pillar0.2-nuscenes](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/centerpoint/centerpoint_02pillar_second_secfpn_4x8_cyclic_20e_nus.py) | 3090<br>bs=4 | 0.858 | 0.441 |
| [centerpoint-pillar0.2-nuscenes](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/centerpoint/centerpoint_02pillar_second_secfpn_4x8_cyclic_20e_nus.py) | 3090<br>bs=8 | 1.681 | 0.832 |
| [centerpoint-voxel0.075-nuscenes](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/centerpoint/centerpoint_0075voxel_second_secfpn_4x8_cyclic_20e_nus.py) | 3090<br>bs=4 | 1.638 | 1.212 |

Score comparison with MMDet3D Original (May Not Precise!!!!! I also made a lot of code revisions besides voxelization and I messed them up)
| conf | training on | score of this version | score by MMDet3D |
|:-------:|:----------------:|:------------------------------:|:-----------------------------:|
| [pointpillars-kitti-3class](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/pointpillars/hv_pointpillars_secfpn_6x8_160e_kitti-3d-3class.py) | 3090*4<br>bs=12 | mAP=61.3359| mAP=59.5 |
| [pointpillars-kitti-car](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/pointpillars/hv_pointpillars_secfpn_6x8_160e_kitti-3d-car.py) | 3090*4<br>bs=12 | mAP=76.7144 | mAP=77.1 |
| [pointpillars-waymo-3class](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/pointpillars/hv_pointpillars_secfpn_sbn_2x16_2x_waymo-3d-3class.py) | 3090*4<br>bs=8 fp16 | L1-mAP=70.04<br>L1-mAPH=63.94<br>L2-mAP=63.62<br>L2-mAPH=58.08| L1-mAP=68.843<br>L1-mAPH=63.318<br>L2-mAP=62.619<br>L2-mAPH=57.573
| [centerpoint-pillar0.2-nuscenes](https://github.com/open-mmlab/mmdetection3d/blob/master/configs/centerpoint/centerpoint_02pillar_second_secfpn_4x8_cyclic_20e_nus.py) | 3090*4<br>bs=8 | mAP=49.43<br>NDS=60.14 | mAP=49.07<br>NDS=59.66