@ZwwWayne 2021-09-02
Hi @zhanggefan ,
Thanks for the efforts. Faster implementation of voxelization is definitely beneficial to the model training and testing with MMDetection3D. For completeness, shall we also leave a comparison of accuracy in this PR? For example, it is better to leave a table to compare the AP of a similar model as a reference, and we may also leave more detailed comments to indicate why it is faster but non-deterministic in the code and documentation.