@zhanggefan 2021-09-03
Hi, @ZwwWayne 
This PR is related to issue #894. Some details of the implementation can be found there.
The score comparison table is added in the PR description now. I may not be precise!!! Because I made many code optimizations(to improve the training speed, but I cannot guarantee that they have no impact on scores) and the voxelization is just one of them.