@ZwwWayne 2021-09-07
> Hi, @ZwwWayne
> This PR is related to issue #894. Some details of the implementation can be found there.
> The score comparison table is added in the PR description now. It may not be precise!!! Because I made many code optimizations(to improve the training speed, but I cannot guarantee that they have no impact on scores) and the voxelization is just one of them.

I see. Can we refer to this PR and issue #894 in the docstring of the module? So that future developers could know the trade-off and discussion. Furthermore, we also welcome you to create PRs for optimization to improve the training speed.