@zhanggefan 2021-09-08
Docstring is revised. Besides, a bug similar to #768 in the previous commit is fixed.